+++ src/client/components/base/Title/Title.css
@@ -73,6 +73,10 @@
   color: white;
 }
 
+.Title__title4--color-white {
+  color: white;
+}
+
 /*
     MediaQ
 */
