+++ src/client/components/hero/Hero.component.js
@@ -2,19 +2,12 @@ import React from 'react';
 import {connect} from 'react-redux';
 import Swiper from 'react-id-swiper';
 import {isMobile} from 'react-device-detect';
-import {Parallax, Background} from 'react-parallax';
 
-import Icon from '../base/Icon';
-import Title from '../base/Title';
-import Button from '../base/Button';
-
-import {getLeavesMap, tagsToUrlParams} from '../../utils/taxonomy';
-
-import {startAnimate} from '../../redux/animate.reducer';
+/*** templates ***/
+import SearchbarTemplate from './templates/Searchbar.template.js';
+import InfoTemplate from './templates/Info.template.js';
 
 import './Hero.css';
-
-const leavesMap = getLeavesMap();
 
 const params = {
   mousewheel: false,
@@ -22,10 +15,10 @@ const params = {
   noSwiping: !isMobile,
   speed: 1000,
   loop: true,
-  autoplay: {
-    delay: 10000,
-    disableOnInteraction: false
-  },
+  // autoplay: {
+  //   delay: 10000,
+  //   disableOnInteraction: false
+  // },
   slidesPerGroup: 1,
   effect: 'fade',
   pagination: {
@@ -36,100 +29,30 @@ const params = {
 };
 
 export class Hero extends React.Component {
-  inRange(tag) {
-    const aCards = Object.values(this.props.aCards);
-    let inRange = false;
-    aCards.forEach(card => {
-      if (card.range && card.range.includes(tag)) {
-        inRange = true;
-        return;
+  template(hero, idx) {
+    switch (hero.template) {
+      case 'searchbar': {
+        return (
+          <SearchbarTemplate
+            className="swiper-slide"
+            key={`${hero.title}-${idx}`}
+            hero={hero}
+          />
+        );
       }
-    });
-    return inRange;
-  }
+      case 'info': {
+        return (
+          <InfoTemplate
+            className="swiper-slide"
+            key={`${hero.title}-${idx}`}
+            hero={hero}
+          />
+        );
+      }
 
-  buildUrl(tags) {
-    return `/find?tags=${tagsToUrlParams(tags)}`;
-  }
-
-  buildTags(tags) {
-    return tags.map(tag => {
-      const cur = this.props.animate[`tag-${tag}`];
-      const hideTags = cur && cur.animating ? 'tags-hidden' : '';
-
-      return (
-        <span
-          key={tag}
-          ref={e => (this[tag] = e)}
-          className="searchbar-tags mt-2 mt-lg-0 d-inline-block"
-          onClick={e => {
-            e.preventDefault();
-            e.stopPropagation();
-
-            if (window.innerWidth < 992) {
-              this.props.redirect(this.buildUrl([tag]));
-              return;
-            }
-
-            this.toggleSearchbar('open');
-            this.resetSearchbarPlaceholder();
-
-            this.props.startAnimation({
-              name: `tag-${tag}`,
-              component: this.buildAnimationTags([tag]),
-              from: this.getPosition(this[tag]),
-              to: this.getPosition(document.getElementById('selectedFilters')),
-              onEnd: () => this.props.redirect(this.buildUrl([tag]))
-            });
-          }}
-        >
-          <Button size="large" type="term" className={`${hideTags} mr-3 h-100`}>
-            {leavesMap[tag].title}
-          </Button>
-        </span>
-      );
-    });
-  }
-
-  buildAnimationTags(tags) {
-    return tags.map(tag => {
-      return (
-        <span
-          key={tag}
-          className="searchbar-tags searchbar-tags-animate mt-2 mt-lg-0 d-inline-block"
-        >
-          <Button size="large" type="term" className="mr-3 h-100">
-            <span>{leavesMap[tag].title}</span>
-            <Icon className="md-small" name="close" />
-          </Button>
-        </span>
-      );
-    });
-  }
-
-  toggleSearchbar(status) {
-    const searchbar = document.getElementById('topbar');
-    if (status === 'open') {
-      searchbar.classList.remove('searchBar-closed');
-    } else if (status === 'close') {
-      searchbar.classList.add('searchBar-closed');
-    } else {
-      return;
+      default:
+        return <div>{'unknown template'}</div>;
     }
-  }
-
-  resetSearchbarPlaceholder() {
-    const searchbar = document.getElementById('Searchbar__inputfield');
-    if (searchbar) {
-      searchbar.placeholder = '';
-    }
-  }
-
-  getPosition(element) {
-    return {
-      top: element.getBoundingClientRect().top,
-      left: element.getBoundingClientRect().left
-    };
   }
 
   render() {
@@ -147,113 +70,9 @@ export class Hero extends React.Component {
           {!heroesIsLoading &&
             heroes.map((hero, idx) => {
               if (!hero.disabled) {
-                const styles = {
-                  backgroundImage: `url(${hero.img})`
-                };
-
-                const shadow =
-                  hero.color === 'white' ? 'black-shadow' : 'white-shadow';
-
-                const filter =
-                  hero.color === 'white' ? 'darkFilter' : 'lightFilter';
-
-                const url = this.buildUrl(hero.tags);
-
-                const hideTags =
-                  animate.tagsAnimation && animate.tagsAnimation.animating
-                    ? 'tags-hidden'
-                    : '';
-
-                return (
-                  <Parallax
-                    key={`${hero.title}-${idx}`}
-                    strength={250}
-                    renderLayer={() => (
-                      <div className={`hero-filter ${filter}`} />
-                    )}
-                  >
-                    <Background>
-                      <div style={styles} className={`hero-bg-image`} />
-                    </Background>
-                    <div className={`text-center position-relative`}>
-                      <Title
-                        Tag="h1"
-                        type="title1"
-                        variant={
-                          hero.color !== 'white' ? `color-${hero.color}` : null
-                        }
-                        className={`${shadow}`}
-                      >
-                        {hero.title}
-                      </Title>
-
-                      <div className="col-12 col-md-10 col-lg-10 col-xl-8 text-left mr-auto ml-auto pt-3 pt-sm-5">
-                        <Title
-                          Tag="h3"
-                          type="title3"
-                          className={`${shadow} mb-4`}
-                          variant={
-                            hero.color !== 'petroleum'
-                              ? `color-${hero.color}`
-                              : null
-                          }
-                        >
-                          {hero.text}
-                        </Title>
-                        <span
-                          className="searchbar-link"
-                          onClick={() => {
-                            if (window.innerWidth < 992) {
-                              redirect(url);
-                              return;
-                            }
-
-                            this.toggleSearchbar('open');
-                            this.resetSearchbarPlaceholder();
-
-                            startAnimation({
-                              name: 'tagsAnimation',
-                              component: this.buildAnimationTags(hero.tags),
-                              from: this.getPosition(this.searchbar),
-                              to: this.getPosition(
-                                document.getElementById('selectedFilters')
-                              ),
-                              onEnd: () => redirect(url)
-                            });
-                          }}
-                        >
-                          <div className="searchbar p-0 p-sm-2 d-inline-flex d-lg-flex flex-column flex-lg-row justify-content-between">
-                            <div
-                              className={'searchbar-innerWrap d-inline-flex'}
-                            >
-                              <Icon
-                                name="search"
-                                className="md-xlarge align-self-center d-none d-lg-inline-block mr-3 ml-2"
-                              />
-                              <span
-                                className={`${hideTags} d-inline-flex flex-column flex-md-row h-100`}
-                                ref={e => (this.searchbar = e)}
-                              >
-                                {this.buildTags(hero.tags)}
-                              </span>
-                            </div>
-                            <Button
-                              type="primary"
-                              size="large"
-                              className="searchbar-button mt-4 mt-lg-0 align-self-start"
-                              variant={`bgcolor-${hero.btnColor}--color-${
-                                hero.btnTextColor
-                              }`}
-                            >
-                              {hero.btnText}
-                            </Button>
-                          </div>
-                        </span>
-                      </div>
-                    </div>
-                  </Parallax>
-                );
+                return this.template(hero, idx);
               }
+              return null;
             })}
         </Swiper>
       </div>
@@ -264,16 +83,13 @@ export class Hero extends React.Component {
 const mapStateToProps = state => {
   return {
     heroes: state.heroReducer.heroes,
-    heroesIsLoading: state.heroReducer.isLoading,
-    aCards: state.filtercardReducer,
-    animate: state.animateReducer
+    heroesIsLoading: state.heroReducer.isLoading
   };
 };
 
-const mapDispatchToProps = dispatch => ({
-  startAnimation: obj => dispatch(startAnimate(obj)),
-  redirect: path => dispatch({type: 'HISTORY_PUSH', path})
-});
+const mapDispatchToProps = dispatch => {
+  return {};
+};
 
 export default connect(
   mapStateToProps,
