+++ src/client/components/hero/Hero.css
@@ -14,7 +14,7 @@
 .Hero .swiper-container-horizontal > .swiper-pagination-bullets {
   top: auto;
   text-align: center;
-  bottom: 10px;
+  top: 50px;
 }
 
 .Hero .swiper-pagination-bullet {
@@ -41,77 +41,6 @@
   border-radius: 5px;
   width: 100%;
 }
-
-/* Hero tags animation */
-
-.tags-hidden {
-  opacity: 0 !important;
-}
-
-.searchbar-tags-animate:last-child {
-  margin-right: 0;
-}
-
-.searchbar-tags-animate:first-child {
-  margin-left: 0;
-
-  -webkit-animation: firsttag 0.5s forwards, tag 0.5s forwards;
-  animation: firsttag 0.5s forwards, tag 0.5s forwards;
-}
-
-@-webkit-keyframes firsttag {
-  100% {
-    margin-left: 0.5rem;
-  }
-}
-
-.searchbar-tags-animate {
-  height: 50px;
-
-  -webkit-animation: tag 0.5s forwards;
-  animation: tag 0.5s forwards;
-}
-
-@-webkit-keyframes tag {
-  100% {
-    height: 32px;
-  }
-}
-
-.searchbar-tags-animate button {
-  padding: 10px 15px;
-  font-size: 1.3em;
-
-  -webkit-animation: button 0.5s forwards;
-  animation: button 0.5s forwards;
-}
-
-.searchbar-tags-animate button span {
-  vertical-align: middle;
-}
-
-@-webkit-keyframes button {
-  100% {
-    padding: 7px 10px 7px 10px;
-    font-size: 14px;
-  }
-}
-
-.searchbar-tags-animate i {
-  margin-left: -20px;
-  opacity: 0;
-  -webkit-animation: buttonicon 0.5s forwards;
-  animation: buttonicon 0.5s forwards;
-}
-
-@-webkit-keyframes buttonicon {
-  100% {
-    margin-left: 10px;
-    opacity: 1;
-  }
-}
-
-/* Hero tags animation END*/
 
 .Hero .searchbar .searchbar-innerWrap i {
   color: var(--petroleum);
@@ -185,61 +114,6 @@
 .Hero .swiper-pagination-bullet-active {
   opacity: 0.9;
 }
-
-/* force searchbar on landscape */
-
-@media (max-width: 991px) and (max-height: 749px) {
-  .Hero {
-    height: calc(100vh - 80px);
-  }
-  .Hero .hero-bg-image {
-    height: calc(100vh + 80px);
-  }
-}
-
-@media (max-width: 992px) and (orientation: landscape) {
-  .Hero .searchbar {
-    display: flex !important;
-    flex-direction: row !important;
-    background: rgba(255, 255, 255, 0.5) !important;
-    border: 1px solid var(--silver-chalice) !important;
-  }
-  .Hero .searchbar .searchbar-innerWrap {
-    flex-direction: row !important;
-  }
-
-  .Hero .searchbar .searchbar-innerWrap > span {
-    flex-direction: row !important;
-  }
-
-  .Hero .searchbar .searchbar-innerWrap i {
-    display: inline-block !important;
-  }
-  .Hero .searchbar .searchbar-button {
-    border: 1px solid white !important;
-  }
-  .searchbar-button,
-  .searchbar-tags {
-    margin-top: 0 !important;
-  }
-  .Hero .searchbar-link {
-    padding-left: 0 !important;
-  }
-}
-
-@media (max-height: 375px) and (orientation: landscape) {
-  .Hero .searchbar-link {
-    transform: scale(0.65);
-  }
-  .Title__title3 {
-    display: none;
-  }
-  .Hero .searchbar-link {
-    padding-left: 0 !important;
-  }
-}
-
-/* force searchbar on landscape END */
 
 @media (max-width: 991px) {
   .Hero .searchbar {
